remove ANIs and extract extremely common interfaces
