+++ XRE.sln
@@ -83,6 +83,10 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Bootstrapper.FunctionalTest
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.ApplicationHost2", "src\Microsoft.Framework.ApplicationHost2\Microsoft.Framework.ApplicationHost2.kproj", "{F6AA0E08-2A20-4019-9720-588D3A0DAF2C}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Runtime.Hosting.Interfaces", "src\Microsoft.Framework.Runtime.Hosting.Interfaces\Microsoft.Framework.Runtime.Hosting.Interfaces.kproj", "{0619AAF7-0B57-4E00-9922-182A328468F1}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Runtime.Roslyn.Interfaces", "src\Microsoft.Framework.Runtime.Roslyn.Interfaces\Microsoft.Framework.Runtime.Roslyn.Interfaces.kproj", "{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -589,6 +593,46 @@ Global
 		{F6AA0E08-2A20-4019-9720-588D3A0DAF2C}.Release|x64.Build.0 = Release|Any CPU
 		{F6AA0E08-2A20-4019-9720-588D3A0DAF2C}.Release|x86.ActiveCfg = Release|Any CPU
 		{F6AA0E08-2A20-4019-9720-588D3A0DAF2C}.Release|x86.Build.0 = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|Win32.Build.0 = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|x64.Build.0 = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Debug|x86.Build.0 = Debug|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Win32.ActiveCfg = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|Win32.Build.0 = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|x64.ActiveCfg = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|x64.Build.0 = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|x86.ActiveCfg = Release|Any CPU
+		{0619AAF7-0B57-4E00-9922-182A328468F1}.Release|x86.Build.0 = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|Win32.Build.0 = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|x64.Build.0 = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Debug|x86.Build.0 = Debug|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Win32.ActiveCfg = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|Win32.Build.0 = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|x64.ActiveCfg = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|x64.Build.0 = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|x86.ActiveCfg = Release|Any CPU
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -625,5 +669,7 @@ Global
 		{FE569E9D-AAAE-43CE-BE29-2C2197C0DF8F} = {13ED5001-B871-4BC3-8499-29607F596C7C}
 		{3F59A0B7-745D-4ADA-A913-6A1AD1A149C2} = {C43EE429-DE10-4906-BB09-54E6A080948A}
 		{F6AA0E08-2A20-4019-9720-588D3A0DAF2C} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
+		{0619AAF7-0B57-4E00-9922-182A328468F1} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
+		{F8DC9AE8-54E1-443C-BF04-60AB0D169E3A} = {AF391791-F4B7-41AC-8F08-9485DAC543C5}
 	EndGlobalSection
 EndGlobal
