+++ src/Microsoft.Framework.DesignTimeHost/project.json
@@ -5,6 +5,7 @@
         "Microsoft.Framework.Runtime": "1.0.0-*",
         "Microsoft.Framework.Runtime.Common": { "version": "1.0.0-*", "type": "build" },
         "Microsoft.Framework.Runtime.Interfaces": { "version": "1.0.0-*", "type": "build" },
+		"Microsoft.Framework.Runtime.Hosting.Interfaces": "1.0.0-*",
         "Microsoft.Framework.Runtime.Roslyn" : "1.0.0-*"
     },
     "compilationOptions" : { "define" : ["TRACE"], "warningsAsErrors": true },
