+++ src/Microsoft.Framework.Runtime.Hosting.Interfaces/CallContextServiceLocator.cs
@@ -3,7 +3,6 @@
 
 namespace Microsoft.Framework.Runtime.Infrastructure
 {
-    [AssemblyNeutral]
     public static class CallContextServiceLocator
     {
         public static IServiceProviderLocator Locator;
