+++ src/Microsoft.Framework.Runtime.Hosting.Interfaces/IApplicationEnvironment.cs
@@ -5,7 +5,6 @@ using System.Runtime.Versioning;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IApplicationEnvironment
     {
         string ApplicationName { get; }
