+++ src/Microsoft.Framework.Runtime.Hosting.Interfaces/IAssemblyLoaderContainer.cs
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis;
+
+using System;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
-    public interface IRoslynMetadataReference : IMetadataReference
+    public interface IAssemblyLoaderContainer
     {
-        MetadataReference MetadataReference { get; }
+        IDisposable AddLoader(IAssemblyLoader loader);
     }
 }
