+++ src/Microsoft.Framework.Runtime.Hosting.Interfaces/IServiceProviderLocator.cs
@@ -1,14 +1,12 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-
 using System;
 
-namespace Microsoft.Framework.Runtime
+namespace Microsoft.Framework.Runtime.Infrastructure
 {
-    [AssemblyNeutral]
-    public interface IAssemblyLoaderContainer
+    public interface IServiceProviderLocator
     {
-        IDisposable AddLoader(IAssemblyLoader loader);
+        IServiceProvider ServiceProvider { get; set; }
     }
 }
