+++ src/Microsoft.Framework.Runtime.Interfaces/IAssemblyLoadContextFactory.cs
@@ -2,7 +2,6 @@
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IAssemblyLoadContextFactory
     {
         IAssemblyLoadContext Create();
