+++ src/Microsoft.Framework.Runtime.Interfaces/ICompilerOptions.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ICompilerOptions
     {
         IEnumerable<string> Defines { get; }
