+++ src/Microsoft.Framework.Runtime.Interfaces/IDiagnosticResult.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IDiagnosticResult
     {
         bool Success { get; }
