+++ src/Microsoft.Framework.Runtime.Interfaces/IFileMonitor.cs
@@ -5,7 +5,6 @@ using System;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IFileMonitor
     {
         event Action<string> OnChanged;
