+++ src/Microsoft.Framework.Runtime.Interfaces/ILibraryExport.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ILibraryExport
     {
         IList<IMetadataReference> MetadataReferences { get; }
