+++ src/Microsoft.Framework.Runtime.Interfaces/ILibraryExportProvider.cs
@@ -5,7 +5,6 @@ using System.Runtime.Versioning;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ILibraryExportProvider
     {
         ILibraryExport GetLibraryExport(ILibraryKey target);
