+++ src/Microsoft.Framework.Runtime.Interfaces/ILibraryKey.cs
@@ -7,7 +7,6 @@ using System.Runtime.Versioning;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ILibraryKey
     {
         string Name { get; }
