+++ src/Microsoft.Framework.Runtime.Interfaces/ILibraryManager.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ILibraryManager
     {
         ILibraryExport GetLibraryExport(string name);
