+++ src/Microsoft.Framework.Runtime.Interfaces/IMetadataFileReference.cs
@@ -4,7 +4,6 @@
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IMetadataFileReference : IMetadataReference
     {
         string Path { get; }
