+++ src/Microsoft.Framework.Runtime.Interfaces/IMetadataProjectReference.cs
@@ -7,7 +7,6 @@ using System.Reflection;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IMetadataProjectReference : IMetadataReference
     {
         string ProjectPath { get; }
