+++ src/Microsoft.Framework.Runtime.Interfaces/IMetadataReference.cs
@@ -4,7 +4,6 @@
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface IMetadataReference
     {
         string Name { get; }
