+++ src/Microsoft.Framework.Runtime.Interfaces/ISourceFileReference.cs
@@ -3,7 +3,6 @@
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
     public interface ISourceFileReference : ISourceReference
     {
         string Path { get; }
