--- src/Microsoft.Framework.Runtime.Interfaces/Infrastructure/IServiceProviderLocator.cs
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-
-namespace Microsoft.Framework.Runtime.Infrastructure
-{
-    [AssemblyNeutral]
-    public interface IServiceProviderLocator
-    {
-        IServiceProvider ServiceProvider { get; set; }
-    }
-}
