+++ src/Microsoft.Framework.Runtime.Loader/project.json
@@ -2,7 +2,7 @@
     "version": "1.0.0-*",
     "description": "ASP.NET 5 runtime infrastructure for assembly load contexts.",
     "dependencies": {
-        "Microsoft.Framework.Runtime.Interfaces": { "version": "1.0.0-*", "type": "build" }
+        "Microsoft.Framework.Runtime.Hosting.Interfaces": { "version": "1.0.0-*", "type": "build" }
     },
 
     "frameworks": {
