+++ src/Microsoft.Framework.Runtime.Roslyn.Interfaces/IRoslynMetadataReference.cs
@@ -1,13 +1,12 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
+using Microsoft.CodeAnalysis;
 
 namespace Microsoft.Framework.Runtime
 {
-    [AssemblyNeutral]
-    [AttributeUsage(AttributeTargets.All, Inherited = false, AllowMultiple = true)]
-    public sealed class AssemblyNeutralAttribute : Attribute
+    public interface IRoslynMetadataReference : IMetadataReference
     {
+        MetadataReference MetadataReference { get; }
     }
 }
