+++ src/Microsoft.Framework.Runtime.Roslyn.Interfaces/project.json
@@ -1,0 +1,19 @@
+﻿{
+    "version": "1.0.0-*",
+    "dependencies": {
+		"Microsoft.Framework.Runtime.Interfaces": "1.0.0-*",
+        "Microsoft.CodeAnalysis.CSharp": "1.0.0-rc1-*"
+    },
+
+    "frameworks" : {
+        "aspnet50" : { 
+            "dependencies": {
+            }
+        },
+        "aspnetcore50" : { 
+            "dependencies": {
+                "System.Runtime": "4.0.20-beta-22523"
+            }
+        }
+    }
+}
