+++ src/Microsoft.Framework.Runtime.Roslyn/project.json
@@ -7,7 +7,8 @@
         "Microsoft.Framework.Runtime": "1.0.0-*",
         "Microsoft.Framework.Runtime.Common": { "version": "1.0.0-*", "type": "build" },
         "Microsoft.Framework.Runtime.Interfaces": { "version": "1.0.0-*", "type": "build" },
-        "Microsoft.Framework.Runtime.Roslyn.Common": { "version": "1.0.0-*", "type": "build" }
+        "Microsoft.Framework.Runtime.Roslyn.Common": { "version": "1.0.0-*", "type": "build" },
+		"Microsoft.Framework.Runtime.Roslyn.Interfaces": "1.0.0-*"
     },
     "frameworks": {
         "aspnet50": {
