+++ src/kre.host/project.json
@@ -3,7 +3,7 @@
     "compilationOptions": { "define": [ "TRACE" ], "warningsAsErrors": true },
     "dependencies": {
         "Microsoft.Framework.Runtime.Common": { "version": "1.0.0-*", "type": "build" },
-        "Microsoft.Framework.Runtime.Interfaces": { "version": "1.0.0-*", "type": "build" },
+        "Microsoft.Framework.Runtime.Hosting.Interfaces": "1.0.0-*",
         "Microsoft.Framework.Runtime.Loader": "1.0.0-*"
     },
     "frameworks": {
